add support for otlp/http with protobuf
