+++ route/route_test.go
@@ -3,8 +3,6 @@ package route
 import (
 	"bytes"
 	"compress/gzip"
-	"context"
-	"encoding/hex"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -18,13 +16,9 @@ import (
 	"github.com/facebookgo/inject"
 	"github.com/honeycombio/refinery/collect"
 	"github.com/honeycombio/refinery/config"
-	collectortrace "github.com/honeycombio/refinery/internal/opentelemetry-proto-gen/collector/trace/v1"
-	common "github.com/honeycombio/refinery/internal/opentelemetry-proto-gen/common/v1"
-	trace "github.com/honeycombio/refinery/internal/opentelemetry-proto-gen/trace/v1"
 	"github.com/honeycombio/refinery/logger"
 	"github.com/honeycombio/refinery/metrics"
 	"github.com/honeycombio/refinery/transmit"
-	"github.com/stretchr/testify/assert"
 
 	"github.com/gorilla/mux"
 	"github.com/honeycombio/refinery/sharder"
@@ -399,160 +393,6 @@ func TestDebugTrace(t *testing.T) {
 	}
 }
 
-func TestOTLPHandler(t *testing.T) {
-	md := metadata.New(map[string]string{"x-honeycomb-team": "meow", "x-honeycomb-dataset": "ds"})
-	ctx := metadata.NewIncomingContext(context.Background(), md)
-
-	mockMetrics := metrics.MockMetrics{}
-	mockMetrics.Start()
-	mockTransmission := &transmit.MockTransmission{}
-	mockTransmission.Start()
-	router := &Router{
-		Config:               &config.MockConfig{},
-		Metrics:              &mockMetrics,
-		UpstreamTransmission: mockTransmission,
-		iopLogger: iopLogger{
-			Logger:         &logger.NullLogger{},
-			incomingOrPeer: "incoming",
-		},
-	}
-
-	conf := &config.MockConfig{
-		GetSendDelayVal:    0,
-		GetTraceTimeoutVal: 60 * time.Second,
-		GetSamplerTypeVal:  &config.DeterministicSamplerConfig{SampleRate: 1},
-		SendTickerVal:      2 * time.Millisecond,
-		GetInMemoryCollectorCacheCapacityVal: config.InMemoryCollectorCacheCapacity{
-			CacheCapacity: 100,
-			MaxAlloc:      100,
-		},
-	}
-
-	t.Run("span with status", func(t *testing.T) {
-		req := &collectortrace.ExportTraceServiceRequest{
-			ResourceSpans: []*trace.ResourceSpans{{
-				InstrumentationLibrarySpans: []*trace.InstrumentationLibrarySpans{{
-					Spans: helperOTLPRequestSpansWithStatus(),
-				}},
-			}},
-		}
-		_, err := router.Export(ctx, req)
-		if err != nil {
-			t.Errorf(`Unexpected error: %s`, err)
-		}
-		assert.Equal(t, 2, len(mockTransmission.Events))
-		mockTransmission.Flush()
-	})
-
-	t.Run("span without status", func(t *testing.T) {
-		req := &collectortrace.ExportTraceServiceRequest{
-			ResourceSpans: []*trace.ResourceSpans{{
-				InstrumentationLibrarySpans: []*trace.InstrumentationLibrarySpans{{
-					Spans: helperOTLPRequestSpansWithoutStatus(),
-				}},
-			}},
-		}
-		_, err := router.Export(ctx, req)
-		if err != nil {
-			t.Errorf(`Unexpected error: %s`, err)
-		}
-		assert.Equal(t, 2, len(mockTransmission.Events))
-		mockTransmission.Flush()
-	})
-
-	// TODO: (MG) figuure out how we can test JSON created from OTLP requests
-	// Below is example, but requires significant usage of collector, sampler, conf, etc
-	t.Run("creates events for span events", func(t *testing.T) {
-		t.Skip("need additional work to support inspecting outbound JSON")
-
-		traceID := []byte{0, 0, 0, 0, 1}
-		spanID := []byte{1, 0, 0, 0, 0}
-		req := &collectortrace.ExportTraceServiceRequest{
-			ResourceSpans: []*trace.ResourceSpans{{
-				InstrumentationLibrarySpans: []*trace.InstrumentationLibrarySpans{{
-					Spans: []*trace.Span{{
-						TraceId: traceID,
-						SpanId:  spanID,
-						Name:    "span_with_event",
-						Events: []*trace.Span_Event{{
-							TimeUnixNano: 12345,
-							Name:         "span_link",
-							Attributes: []*common.KeyValue{{
-								Key: "event_attr_key", Value: &common.AnyValue{Value: &common.AnyValue_StringValue{StringValue: "event_attr_val"}},
-							}},
-						}},
-					}},
-				}},
-			}},
-		}
-		_, err := router.Export(ctx, req)
-		if err != nil {
-			t.Errorf(`Unexpected error: %s`, err)
-		}
-
-		time.Sleep(conf.SendTickerVal * 2)
-
-		mockTransmission.Mux.Lock()
-		assert.Equal(t, 2, len(mockTransmission.Events))
-
-		spanEvent := mockTransmission.Events[0]
-		// assert.Equal(t, time.Unix(0, int64(12345)).UTC(), spanEvent.Timestamp)
-		assert.Equal(t, bytesToTraceID(traceID), spanEvent.Data["trace.trace_id"])
-		assert.Equal(t, hex.EncodeToString(spanID), spanEvent.Data["trace.span_id"])
-		assert.Equal(t, "span_link", spanEvent.Data["span.name"])
-		assert.Equal(t, "span_with_event", spanEvent.Data["parent.name"])
-		assert.Equal(t, "span_event", spanEvent.Data["meta.annotation_type"])
-		assert.Equal(t, "event_attr_key", spanEvent.Data["event_attr_val"])
-		mockTransmission.Mux.Unlock()
-		mockTransmission.Flush()
-	})
-
-	t.Run("creates events for span links", func(t *testing.T) {
-		t.Skip("need additional work to support inspecting outbound JSON")
-
-		traceID := []byte{0, 0, 0, 0, 1}
-		spanID := []byte{1, 0, 0, 0, 0}
-		linkTraceID := []byte{0, 0, 0, 0, 2}
-		linkSpanID := []byte{2, 0, 0, 0, 0}
-
-		req := &collectortrace.ExportTraceServiceRequest{
-			ResourceSpans: []*trace.ResourceSpans{{
-				InstrumentationLibrarySpans: []*trace.InstrumentationLibrarySpans{{
-					Spans: []*trace.Span{{
-						Name:    "span_with_link",
-						TraceId: traceID,
-						SpanId:  spanID,
-						Links: []*trace.Span_Link{{
-							TraceId:    traceID,
-							SpanId:     spanID,
-							TraceState: "link_trace_state",
-							Attributes: []*common.KeyValue{{
-								Key: "link_attr_key", Value: &common.AnyValue{Value: &common.AnyValue_StringValue{StringValue: "link_attr_val"}},
-							}},
-						}},
-					}},
-				}},
-			}},
-		}
-		_, err := router.Export(ctx, req)
-		if err != nil {
-			t.Errorf(`Unexpected error: %s`, err)
-		}
-
-		time.Sleep(conf.SendTickerVal * 2)
-		assert.Equal(t, 2, len(mockTransmission.Events))
-
-		spanLink := mockTransmission.Events[1]
-		assert.Equal(t, bytesToTraceID(traceID), spanLink.Data["trace.trace_id"])
-		assert.Equal(t, hex.EncodeToString(spanID), spanLink.Data["trace.span_id"])
-		assert.Equal(t, bytesToTraceID(linkTraceID), spanLink.Data["trace.link.trace_id"])
-		assert.Equal(t, hex.EncodeToString(linkSpanID), spanLink.Data["trace.link.span_id"])
-		assert.Equal(t, "link", spanLink.Data["meta.annotation_type"])
-		assert.Equal(t, "link_attr_val", spanLink.Data["link_attr_key"])
-		mockTransmission.Flush()
-	})
-}
-
 func TestDependencyInjection(t *testing.T) {
 	var g inject.Graph
 	err := g.Provide(
@@ -572,51 +412,6 @@ func TestDependencyInjection(t *testing.T) {
 	}
 	if err := g.Populate(); err != nil {
 		t.Error(err)
-	}
-}
-
-func helperOTLPRequestSpansWithoutStatus() []*trace.Span {
-	now := time.Now()
-	return []*trace.Span{
-		{
-			StartTimeUnixNano: uint64(now.UnixNano()),
-			Events: []*trace.Span_Event{
-				{
-					TimeUnixNano: uint64(now.UnixNano()),
-					Attributes: []*common.KeyValue{
-						{
-							Key: "attribute_key",
-							Value: &common.AnyValue{
-								Value: &common.AnyValue_StringValue{StringValue: "attribute_value"},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func helperOTLPRequestSpansWithStatus() []*trace.Span {
-	now := time.Now()
-	return []*trace.Span{
-		{
-			StartTimeUnixNano: uint64(now.UnixNano()),
-			Events: []*trace.Span_Event{
-				{
-					TimeUnixNano: uint64(now.UnixNano()),
-					Attributes: []*common.KeyValue{
-						{
-							Key: "attribute_key",
-							Value: &common.AnyValue{
-								Value: &common.AnyValue_StringValue{StringValue: "attribute_value"},
-							},
-						},
-					},
-				},
-			},
-			Status: &trace.Status{Code: trace.Status_STATUS_CODE_OK},
-		},
 	}
 }
